liste dropdown liste, bootstrap-icons , apiStudent in environment.ts

+++ src/app/GestionStudent/Components/students/students.component.ts
@@ -1,9 +1,10 @@
-import { HttpClient } from '@angular/common/http';
-import { Component } from '@angular/core';
-import { Observable } from 'rxjs';
-import { OnInit } from '@angular/core';
-import { StudentServicesService } from '../../Services/student-services.service';
-import { Student1 } from '../../Models/student1';
+import {HttpClient} from '@angular/common/http';
+import {Component} from '@angular/core';
+import {Observable} from 'rxjs';
+import {OnInit} from '@angular/core';
+import {StudentServicesService} from '../../Services/student-services.service';
+import {Student1} from '../../Models/student1';
+
 @Component({
   selector: 'app-students',
   templateUrl: './students.component.html',
@@ -11,30 +12,35 @@ import { Student1 } from '../../Models/student1';
 })
 export class StudentsComponent implements OnInit {
 
-  students: Student1[] =[];
-  constructor(private studentService:StudentServicesService ){}
+  students: Student1[] = [];
+
+  constructor(private studentService: StudentServicesService) {
+  }
+
   ngOnInit(): void {
     this.getStudents();
   }
-  private getStudents(){
-    this.studentService.getStudentsList().subscribe(data=>{
-      this.students=data;
+
+  private getStudents() {
+    this.studentService.getStudentsList().subscribe(data => {
+      this.students = data;
     });
   }
 
   studentUpdated = {
-    nom:"",
-    prenom:"",
-    email:"",
-    adresse:"",
-    phone:"",
-    cin:"",
-    cne:"",
-    apogee:"",
-    lieuNaissance:"",
+    nom: "",
+    prenom: "",
+    email: "",
+    adresse: "",
+    phone: "",
+    cin: "",
+    cne: "",
+    apogee: "",
+    lieuNaissance: "",
   }
+
   updateStudent(): void {
-    this.studentService.updateStudent(this.studentUpdated.apogee ,this.studentUpdated).subscribe({
+    this.studentService.updateStudent(this.studentUpdated.apogee, this.studentUpdated).subscribe({
       next: () => {
         this.getStudents();
       },
@@ -44,19 +50,22 @@ export class StudentsComponent implements OnInit {
     });
   }
 
-  edit( student: any) {
+  edit(student: any) {
     this.studentUpdated = student;
   }
 
   deleteStudent(apogee: number): void {
-    this.studentService.deleteStudent(apogee).subscribe({
-      next: () => {
-        this.getStudents();
-      },
-      error: (error) => {
-        console.log(error);
-      }
-    });
-  }
+    const confirmed = confirm('Vous voulez supprimer cet etudiant ?');
+    if (confirmed) {
+      this.studentService.deleteStudent(apogee).subscribe({
+        next: value => {
+          alert('etudiant supprimé !!');
+          this.getStudents();
+        },
+        error: console.log
+      });
 
+
+    }
+  }
 }
